refactor(analysis-types): split git DTOs
